backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Network & System
psutil==5.9.8
netifaces==0.11.0
python-nmap==0.7.1

# Docker Integration
docker==7.1.0
requests>=2.32.0
urllib3>=2.0.0

# HTTP Client (for Nginx Proxy Manager API)
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
